refactor: move api operations to one controller
